// local-hobbies-frontend/src/types/availability.ts
export enum AvailabilityType {
    OneTimeAvailability = 'One-time availability',
    RecurringAvailability = 'Recurring availability',
    AvailabilityException = 'Availability exception',
}
// ...
export interface AvailabilityIntervalResponse {
    sourceType: AvailabilityType;
    sourceId: string;
    location: GeoJsonPoint;
    /** ISO 8601 UTC timestamp */
    start: string;
    /** ISO 8601 UTC timestamp */
    end: string;
}

export interface AvailabilityOverlapResponse {
    distanceKilometers: number;
    /** ISO 8601 UTC timestamp */
    start: string;
    /** ISO 8601 UTC timestamp */
    end: string;
}